+++ libs/PayU_SDK_Lite.xcframework/ios-arm64_i386_x86_64-simulator/PayU_SDK_Lite.framework/Headers/PUApplePayTransaction.h
@@ -6,43 +6,12 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-typedef NS_ENUM(NSUInteger, PUCurrencyCode) {
-    PUCurrencyCodePLN,
-    PUCurrencyCodeCZK,
-    PUCurrencyCodeEUR,
-    PUCurrencyCodeUSD,
-    PUCurrencyCodeGBP,
-    PUCurrencyCodeRON,
-    PUCurrencyCodeHUF,
-    PUCurrencyCodeHRK,
-    PUCurrencyCodeSEK,
-    PUCurrencyCodeNOK,
-    PUCurrencyCodeDKK,
-};
-
-typedef NS_ENUM(NSUInteger, PUCountryCode) {
-    PUCountryCodePL,
-    PUCountryCodeCZ,
-    PUCountryCodeDE,
-    PUCountryCodeUS,
-    PUCountryCodeGB,
-    PUCountryCodeRO,
-    PUCountryCodeHU,
-    PUCountryCodeHR,
-    PUCountryCodeSE,
-    PUCountryCodeNO,
-    PUCountryCodeDK,
-    PUCountryCodeLT,
-    PUCountryCodeLV,
-    PUCountryCodeSK,
-};
-
 @interface PUApplePayTransaction : NSObject
 - (instancetype)initWithMerchantIdentifier:(NSString *)merchantIdentifier
-                              currencyCode:(PUCurrencyCode)currencyCode
-                               countryCode:(PUCountryCode)countryCode
+                              currencyCode:(NSString *)currencyCode
+                               countryCode:(NSString *)countryCode
                        contactEmailAddress:(NSString *)contactEmailAddress
-                   paymentItemDescription:(NSString *)paymentItemDescription
+                    paymentItemDescription:(NSString *)paymentItemDescription
                                     amount:(NSDecimalNumber *)amount;
 @end